Implemented the DELETE endpoint in the books controller. Modified the logging statements of the DELETE endpoint of the authors controller to conform with the logging standard used within the books controller. Corrected a bug in the book data entity model and data transfer objects changing the price property type from double to decimal so that it conforms with the "money" data type in the database.

+++ DTOs/BookDTO.cs
@@ -40,7 +40,7 @@ namespace BookStore_API.DTOs
         /// <summary>
         /// Gets or sets the price of this book.
         /// </summary>
-        public double? Price { get; set; }
+        public decimal? Price { get; set; }
 
         /// <summary>
         /// Gets or sets the unique identifier of the author of this book.
@@ -89,7 +89,7 @@ namespace BookStore_API.DTOs
         /// <summary>
         /// Gets or sets the price of this book.
         /// </summary>
-        public double? Price { get; set; }
+        public decimal? Price { get; set; }
 
         /// <summary>
         /// Gets or sets the unique identifier of the author of this book.
@@ -134,6 +134,6 @@ namespace BookStore_API.DTOs
         /// <summary>
         /// Gets or sets the price of this book.
         /// </summary>
-        public double? Price { get; set; }
+        public decimal? Price { get; set; }
     }
 }
